FactoredRandomNumbers

Description

Module for accessing functions based on Kalai's Algorithm

Documentation

genARandomPreFactoredNumberLTEn :: Integer -> IO (Either Text (Integer, [Integer]))

This is the Entry Function with a Integer bound Provide an integer input and it should generate a tuple of a number less than the input integer and its prime factors

preFactoredNumOfBitSize :: Integer -> IO (Either Text (Integer, [Integer]))

Takes an Integer as a Bitsize value to operate on range [2 ^ y, 2 ^ y + 1 - 1] Provide an integer input and it should generate a tuple of a number in the range [2^y, 2^y+1 -1] and its prime factors

preFactoredNumOfBitSizePar :: Integer -> IO (Either Text (Integer, [Integer]))

Takes an Integer for Bitsize value to operate on range [2 ^ y, 2 ^ y + 1 - 1]. This function leverages parallel execution Provide an integer input and it should generate a tuple of a number in the range [2^y, 2^y+1 -1] and its prime factors In the event that the parallel call fails and return Nothing, a recovery through a non-parallel call is attempted.