Lib

Description

Module for accessing this math function

Documentation

genARandomPreFactoredNumberLTEn' :: Int -> IO (Either Text (Int, [Int]))

This is the Entry Function. Provide an integer input and it should generate a tuple of a number less than the integer i/p and its prime factors

getRndMInt :: (Int, Int) -> IO Int

Get a Random Integer with uniform probability in the range [1,n]

makeList :: Int -> IO [Int]

Provided an Int, creates a sequence of random integers LTE n decreasing possibly with multiples ending at single 1

getName :: IO String

testMap :: [Bool]

testSequence :: [Bool]

testLiftM2 :: Bool

test1LiftM2 :: IO String

testfMap :: Maybe Int

test2liftM2 :: Maybe Int