ShortCircuit

Description

Module for nifty functions leveraging lazy evaluation

Type classes

class HasFalse a where

Types with a defined false value.

Methods

false :: a

A false value for this type. If Shortcircuit a holds, isTrue false == False must hold.

class HasTrue a where

Types with a defined true value.

Methods

true :: a

A true value for this type. If Shortcircuit a holds, isTrue true == True must hold.

class Shortcircuit a where

Types that support short circuits.

Methods

isTrue :: a -> Bool

Whether the value is true-like (i.e. not false-like).

Short circuit tests

isFalse :: Shortcircuit a => a -> Bool

Whether the value is false-like (i.e. not true-like).

if' :: Shortcircuit a => a -> b -> b -> b

if then else generalised to Shortcircuit.

unless' :: Shortcircuit a => a -> b -> b -> b

The opposite of if'.

(??) :: Shortcircuit a => b -> b -> a -> b infix 2

Like if', but with different argument order, allowing infix use.

(||) :: Shortcircuit a => a -> a -> a infix 2

|| generalised to Shortcircuit.

(&&) :: Shortcircuit a => a -> a -> a infix 2

&& generalised to Shortcircuit.

firstTrueOf :: (Shortcircuit a, HasFalse a) => [a] -> a

Returns the first true-ish value from a list, or false.

lastFalseOf :: (Shortcircuit a, HasTrue a) => [a] -> a

Returns the last false-ish value from a list, or true.

firstFalseOf :: (Shortcircuit a, HasTrue a) => [a] -> a

Returns the first False-sh value of a provided list

Monadic short circuits

orM :: (Monad m, Shortcircuit a) => m a -> m a -> m a

Short-circuit two actions, performing the second only if the first returned a false-ish value.

andM :: (Monad m, Shortcircuit a) => m a -> m a -> m a

Short-circuit two actions, performing the second only if the first returned a true-ish value.

firstTrueOfM :: (Monad m, Shortcircuit a, HasFalse a) => [m a] -> m a

Short-circuit a list of actions, performing only until a true-ish value is found, or the list exhausted.

lastFalseOfM :: (Monad m, Shortcircuit a, HasTrue a) => [m a] -> m a

Short-circuit a list of actions, performing only until a false-ish value is found, or the list exhausted.

firstFalseOfM :: (Monad m, Shortcircuit a, HasTrue a) => [m a] -> m a

Returns the first False-ish value from a monadic list