Lib

Description

Module for accessing this math function

Documentation

genARandomPreFactoredNumberLEn :: Int -> IO (Either Text (Int, [Int]))

This is the Entry Function. Provide an integer input and it should generate a tuple of a number less than the integer i/p and its factors

createSeq' :: Int -> [Int]

Creates a sequence of prime ints below the input integer. The input has to be positive int. The output will then be a list of primes (of positive ints).